Various: warning fixes

+++ modules/eigenutils/src/processors/testmatrix.cpp
@@ -50,10 +50,10 @@ TestMatrix::TestMatrix()
 void TestMatrix::process() {
 
     auto m = std::make_shared<Eigen::MatrixXf>(size_.get(), size_.get());
-    for (int y = 0; y < size_.get(); y++) {
+    for (size_t y = 0; y < size_.get(); y++) {
         float Y = y / static_cast<float>(size_.get() - 1);
         Y *= Y * 0.5f;
-        for (int x = 0; x < size_.get(); x++) {
+        for (size_t x = 0; x < size_.get(); x++) {
             float X = x / static_cast<float>(size_.get() - 1);
             X *= X;
             (*m)(y, x) = glm::mix(Y, 1.0f, X);
